requirements.txt
# FastAPI & Server
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
psycopg2
psycopg2-binary
alembic

# Pydantic
pydantic
pydantic-settings
pydantic[email]

# Security & Auth
python-jose[cryptography]
passlib[bcrypt]
bcrypt

# Google Gemini
google-genai

# Image Processing
Pillow

# Scheduler
APScheduler

# Utils
python-dotenv

# Redis (pour cache & tasks avancées)
redis

# Testing
pytest
pytest-asyncio
httpx

# Caching inexistant
